Remove Join Node
Its implementation was the same as Beta Node, so I've asked why. ChatGPT
insisted to provide the same code, I've asked for a description of all
the classes involved and here it is the result:

## Class List Update

Here is an updated list of the classes needed for the Ruleur gem:

1. `RuleSet`: A class that represents a collection of rules.
2. `Network`: A class that represents the Rete algorithm network.
3. `BetaNode`: A class that represents a beta node in the Rete algorithm.
4. `AlphaNode`: A class that represents an alpha node in the Rete algorithm.
5. `Rule`: A class that represents a business rule.
6. `Fact`: A class that represents a piece of information that is relevant to the BRMS.
7. `Version`: A class that represents the version of the gem.
8. `WorkingMemory`: A class that represents the working memory of the BRMS.
9. `Action` (Optional): A class that represents the actions to be performed when a rule is matched.

The `JoinNode` class is not needed as it is the same as the `BetaNode` class.